fix: CRITICAL - Type mismatch in min() call in MQTT backoff calculation
ERROR FOUND DURING COMPILATION:
  File: firmware/tank-sensor/src/main.cpp
  Line: 41
  Issue: min() requires both arguments to be same type

ORIGINAL CODE (FAILED TO COMPILE):
  mqttBackoffMs = (uint16_t)min(mqttBackoffMs * 2, 60000UL);
  Problem: mqttBackoffMs*2 is int, 60000UL is unsigned long - type mismatch

FIXED CODE:
  mqttBackoffMs = (uint16_t)min((uint32_t)(mqttBackoffMs * 2), 60000UL);
  Solution: Cast both args to uint32_t for std::min() compatibility

COMPILATION RESULT:
  ✅ [SUCCESS] Took 9.48 seconds
  ✅ RAM: 15.7% (51572 / 327680 bytes)
  ✅ Flash: 83.3% (1528743 / 1835008 bytes)

CRITICAL LESSON LEARNED:
  - This error was MISSED by 9-angle code review
  - Code review MUST include compilation verification
  - Static analysis alone is insufficient
  - Always run `pio run` before declaring code ready

This fix is required for production deployment.

+++ firmware/tank-sensor/src/main.cpp
@@ -38,7 +38,7 @@ static void mqttEnsureConnected() {
         if (now - lastMqttAttempt >= mqttBackoffMs) {
             if (!mqttClient.connect(config.d.node_id)) {
                 // Failed: increase backoff (double, capped at 60s)
-                mqttBackoffMs = (uint16_t)min(mqttBackoffMs * 2, 60000UL);
+                mqttBackoffMs = (uint16_t)min((uint32_t)(mqttBackoffMs * 2), 60000UL);
             } else {
                 // Connected: reset backoff
                 mqttBackoffMs = 1000;
